+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/feeds/DraftEventsFeedFilter.kt
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.ui.feeds
+
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.Note
+import com.vitorpamplona.amethyst.commons.model.cache.ICacheProvider
+import com.vitorpamplona.quartz.nip37Drafts.DraftWrapEvent
+
+class DraftEventsFeedFilter(
+    val account: IAccount,
+    val cache: ICacheProvider,
+) : AdditiveFeedFilter<Note>() {
+    override fun feedKey(): String = account.userProfile().pubkeyHex
+
+    override fun applyFilter(newItems: Set<Note>): Set<Note> =
+        newItems.filterTo(HashSet()) {
+            acceptableEvent(it)
+        }
+
+    override fun feed(): List<Note> {
+        val drafts =
+            cache.filterAddressablesIntoSet(DraftWrapEvent.KIND, account.userProfile().pubkeyHex) { _, note ->
+                acceptableEvent(note)
+            }
+
+        return sort(drafts)
+    }
+
+    fun acceptableEvent(it: Note): Boolean {
+        val noteEvent = it.event
+        return noteEvent is DraftWrapEvent && noteEvent.pubKey == account.userProfile().pubkeyHex
+    }
+
+    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
+}
